requirements.txt
Flask==3.1.0
requests
google-cloud-firestore